Backend/pyproject.toml
[project]
name = "nextstep-backend"
version = "0.1.0"
description = "NextStep AI Math Tutor Backend"
authors = [
    {name = "Ayush Sharma"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "openai (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic (>=2.13.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
